Add files via upload
Find command doesn't seem to be working. This is likely due to a strange error I'm having where the line of the file I read in has a strange format where spaces are put between all letters, Ex. c o m m i t, even though that's not how it's laid out in the file. I've never seen this happen before and it's really odd. .strip() doesn't remove this either.